pyproject.toml
[project]
name = "llm-client"
version = "0.1.0"
description = "LLM Client for One-API based on OpenAI standard"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "httpx[socks]>=0.27.0",
]

[tool.hatch.build.targets.wheel]
packages = ["llm"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
